+++ test/Unit/FieldDefinitionTest.php
@@ -1,0 +1,52 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Copyright (c) 2020 Andreas Möller
+ *
+ * For the full copyright and license information, please view
+ * the LICENSE.md file that was distributed with this source code.
+ *
+ * @see https://github.com/ergebnis/factory-bot
+ */
+
+namespace Ergebnis\FactoryBot\Test\Unit;
+
+use Ergebnis\FactoryBot\Exception;
+use Ergebnis\FactoryBot\FieldDefinition;
+use Ergebnis\Test\Util\Helper;
+use PHPUnit\Framework;
+
+/**
+ * @internal
+ *
+ * @covers \Ergebnis\FactoryBot\FieldDefinition
+ *
+ * @uses \Ergebnis\FactoryBot\Exception\InvalidCount
+ */
+final class FieldDefinitionTest extends Framework\TestCase
+{
+    use Helper;
+
+    /**
+     * @dataProvider \Ergebnis\FactoryBot\Test\DataProvider\NumberProvider::intLessThanOne()
+     *
+     * @param int $count
+     */
+    public function testReferencesThrowsInvalidCountExceptionWhenCountIsLessThanOne(int $count): void
+    {
+        $className = self::class;
+
+        $this->expectException(Exception\InvalidCount::class);
+        $this->expectExceptionMessage(\sprintf(
+            'Count needs to be greater than or equal to 1, but %d is not.',
+            $count
+        ));
+
+        FieldDefinition::references(
+            $className,
+            $count
+        );
+    }
+}