+++ CHANGELOG.md
@@ -27,6 +27,7 @@ For a full diff see [`fa9c564...master`][fa9c564...master].
 * Renamed `EntityDef` to `EntityDefinition` ([#91]), by [@localheinz]
 * Renamed `FieldDef` to `FieldDefinition` ([#92]), by [@localheinz]
 * Turned `$configuration` parameter of `FixtureFactory::defineEntity()` into `$afterCreate`, a `Closure` that will be invoked after object construction ([#101]), by [@localheinz]
+* Started throwing an `InvalidCount` exception instead of a generic `Exception` when an invalid number of entities are requested ([#105]), by [@localheinz]
 
 ### Fixed
 
@@ -49,5 +50,6 @@ For a full diff see [`fa9c564...master`][fa9c564...master].
 [#91]: https://github.com/ergebnis/factory-bot/pull/91
 [#92]: https://github.com/ergebnis/factory-bot/pull/92
 [#101]: https://github.com/ergebnis/factory-bot/pull/101
+[#105]: https://github.com/ergebnis/factory-bot/pull/105
 
 [@localheinz]: https://github.com/localheinz